Refactor: Move app setup to app.js and update server.js
